added-state-history-service
stateHistory can be helpful if you want to goBack to the last route with the exact parameters.
Example: User press the arrow back in the Order List of Products Search List.
Before there was only one route for Order Details view. Now there are two. To solve this problem we just use $stateHistory.goBack() and the service is in charge of guess what was the last route.